pyproject.toml
[tool.poetry]
name = "liberator"
version = "0.1.0"
description = "porject for website"
authors = ["My Name <pradeep94@outlook.com"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2"
djangorestframework = ">=3.9.0,<3.10.0"
react = "^4.3.0"
django-cors-headers = "^3.14.0"
cassandra-driver = "^3.26.0"
django-cassandra-engine = "^1.8.0"

[tool.poetry.dev-dependencies]

[tool.poetry.extras]


